+++ lib/core/Node.ts
@@ -1,11 +1,11 @@
 /**
- * @desc A basic node class for pathfinding.
+ * @description A basic node class for pathfinding.
  * @author Digitsensitive <digit.sensitivee@gmail.com>
- * @copyright 2017-2019 Digitsensitive
+ * @copyright 2017 - 2020 Digitsensitive
  * @license {@link https://opensource.org/licenses/MIT|MIT License}
  */
 
-import { INodeConstructor, IPoint } from '../interfaces/astar-interfaces';
+import { INodeConstructor, IPoint } from '../interfaces/astar.interfaces';
 
 export class Node {
   //  General properties