+++ lib/interfaces/astar.interfaces.ts
@@ -1,4 +1,11 @@
-import { Heuristic } from '../types/astar-types';
+/**
+ * @description Interfaces used for astar-typescript
+ * @author Digitsensitive <digit.sensitivee@gmail.com>
+ * @copyright 2017 - 2020 Digitsensitive
+ * @license {@link https://opensource.org/licenses/MIT|MIT License}
+ */
+
+import { Heuristic } from '../types/astar.types';
 
 export interface IAStarFinderConstructor {
   grid: IGridConstructor;