+++ lib/types/astar-types.ts
@@ -1,1 +1,0 @@
-export type Heuristic = 'Manhatten' | 'Euclidean' | 'Chebyshev' | 'Octile';