+++ lib/types/astar.types.ts
@@ -1,9 +1,8 @@
 /**
- * @description Exports from astar-typescript
+ * @description Types used for astar-typescript
  * @author Digitsensitive <digit.sensitivee@gmail.com>
  * @copyright 2017 - 2020 Digitsensitive
  * @license {@link https://opensource.org/licenses/MIT|MIT License}
  */
 
-export { AStarFinder } from './finders/astar-finder';
-export { Grid } from './core/grid';
+export type Heuristic = 'Manhatten' | 'Euclidean' | 'Chebyshev' | 'Octile';