+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [1.1.4] - 2020-02-21
+### Added
+- Add `example` folder
+
+### Changed
+- Small changes in the `.gitignore` and `.npmignore` files
+- Update `CHANGELOG.md`
+- Small changes
+
 ## [1.1.3] - 2020-02-16
 ### Added
 - Add `tsconfig.json`file